Tests fixed
double (arrays) are used now

+++ Tests/Check_ConfidenceIntervals.h
@@ -2,10 +2,10 @@
 #include <check.h>
 
 START_TEST(test_conf_itvl_same_values) {
-  unsigned int all_nines[5] = {9, 9, 9, 9, 9};
-  unsigned int *conf_eighty = conf_itvl(all_nines, 5, 0.8);
-  unsigned int *conf_hundred = conf_itvl(all_nines, 5, 1.0);
-  unsigned int *conf_point_1 = conf_itvl(all_nines, 5, 0.1);
+  double all_nines[5] = {9, 9, 9, 9, 9};
+  double *conf_eighty = conf_itvl(all_nines, 5, 0.8);
+  double *conf_hundred = conf_itvl(all_nines, 5, 1.0);
+  double *conf_point_1 = conf_itvl(all_nines, 5, 0.1);
   ck_assert(*conf_eighty == 9);
   ck_assert(*(conf_eighty + 1) == 9);
   ck_assert(*conf_hundred == 9);
